+++ ptls/frames/bert/datasets/mlm_indexed_dataset.py
@@ -2,8 +2,11 @@ import numpy as np
 import torch
 import random
 
+from ptls.data_load import padded_collate_wo_target
 from ptls.data_load.utils import collate_feature_dict
 from ptls.data_load.feature_dict import FeatureDict
+from ptls.data_load.augmentations import sequence_pair_augmentation
+
 
 class MlmIndexedDataset(torch.utils.data.Dataset):
     """
